server1/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
PyMySQL
kafka-python